fix(skills): address consolidation review changes
