+++ .agents/skills/nemoclaw-worktree-cleanup/SKILL.md
